Add support for babelify and browserify
