+++ index.htm
@@ -18,10 +18,12 @@
         </div>
     </div>
 
-    <script src="node_modules/react/umd/react.production.min.js"></script>
+    <!-- <script src="node_modules/react/umd/react.production.min.js"></script>
     <script src="node_modules/react-dom/umd/react-dom.production.min.js"></script>
     <script src="node_modules/react-router-dom/umd/react-router-dom.min.js"></script>
+    <script src="node_modules/flux/dist/Flux.min.js"></script>
+    <script src="node_modules/flux/dist/FluxUtils.min.js"></script> -->
     <!-- <script src="vendor.js"></script> -->
-    <script src="app.js"></script>
+    <script src="bundle.js"></script>
 </body>
 </html>
