+++ package.json
@@ -3,9 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "babel": "^6.23.0",
-    "babel-cli": "^6.26.0",
-    "babel-preset-react-app": "^3.1.2",
+    "flux": "^3.1.3",
+    "immutable": "^4.0.0-rc.12",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-router-dom": "^5.0.1"
@@ -14,6 +13,14 @@
     "extends": "react-app"
   },
   "scripts": {
-    "babel": "npx babel --watch src --out-dir . --presets react-app/prod "
+    "babel": "npx babel --watch src --out-file app.js --presets react-app/prod ",
+    "bundle": ".\\node_modules\\.bin\\browserify .\\src\\root.js -o bundle.js -t [ babelify --presets [ babel-preset-es2015 babel-preset-react ] ]"
+  },
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babelify": "^8.0.0",
+    "browserify": "^16.5.0"
   }
 }
