+++ src/components/AddTodo.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import AddTodoSimple from './AddTodoSimple';
+
+function AddTodo() {
+  return (
+    <AddTodoSimple></AddTodoSimple>
+  );
+}
+
+export default AddTodo;
