+++ src/components/AddTodoSimple.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function AddTodoSimple() {
+  return (
+    <p>AddTodoSimple</p>
+  );
+}
+
+export default AddTodoSimple;
