+++ src/components/EditTodo.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import EditTodoSimple from './EditTodoSimple';
+
+function EditTodo() {
+  return (
+    <EditTodoSimple></EditTodoSimple>
+  );
+}
+
+export default EditTodo;
