+++ src/components/EditTodoSimple.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function EditTodoSimple() {
+  return (
+    <p>EditTodoSimple</p>
+  );
+}
+
+export default EditTodoSimple;
