+++ src/components/Todo.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import TodoSimple from './TodoSimple';
+
+function Todo() {
+  return (
+    <TodoSimple></TodoSimple>
+  );
+}
+
+export default Todo;
