+++ src/components/TodoList.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import TodoListSimple from './TodoListSimple';
+
+function TodoList() {
+  return (
+    <TodoListSimple></TodoListSimple>
+  );
+}
+
+export default TodoList;
