+++ src/components/TodoListSimple.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Todo from './Todo';
+
+function TodoListSimple() {
+  return (
+    <React.Fragment>
+      <p>TodoListSimple</p>
+      <Todo></Todo>
+    </React.Fragment>
+  );
+}
+
+export default TodoListSimple;
