+++ src/components/TodoSimple.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function TodoSimple() {
+  return (
+    <p>TodoSimple</p>
+  );
+}
+
+export default TodoSimple;
