+++ src/data/TodoActionTypes.js
@@ -1,0 +1,5 @@
+const ActionTypes = {
+  ADD_TODO: 'ADD_TODO',
+};
+
+export default ActionTypes;
