+++ src/data/TodoDispatcher.js
@@ -1,0 +1,3 @@
+import { Dispatcher } from 'flux';
+
+export default new Dispatcher();
