+++ src/data/TodoStore.js
@@ -1,0 +1,27 @@
+import Immutable from 'immutable';
+import { ReduceStore } from 'flux/utils';
+import TodoActionTypes from './TodoActionTypes';
+import TodoDispatcher from './TodoDispatcher';
+
+class TodoStore extends ReduceStore {
+  constructor() {
+    super(TodoDispatcher);
+  }
+
+  getInitialState() {
+    return Immutable.OrderedMap();
+  }
+
+  reduce(state, action) {
+    switch (action.type) {
+      case TodoActionTypes.ADD_TODO:
+        // Do nothing for now, we will add logic here soon!
+        return state;
+
+      default:
+        return state;
+    }
+  }
+}
+
+export default new TodoStore();
