+++ src/pages/AddTodoPage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import AddTodo from '../components/AddTodo';
+
+function AddTodoPage() {
+  return (
+    <React.Fragment>
+      <p>AddTodoPage</p>
+      <AddTodo></AddTodo>
+    </React.Fragment>
+  );
+}
+
+export default AddTodoPage;
