+++ src/pages/AppPage.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { Route, Switch } from 'react-router-dom';
+import TodoListPage from '../pages/TodoListPage';
+import AddTodoPage from '../pages/AddTodoPage';
+import EditTodoPage from '../pages/EditTodoPage';
+import NotFoundPage from '../pages/NotFoundPage';
+
+function AppPage() {
+  return (
+    <React.Fragment>
+      <h1>Todo App</h1>
+      <Switch>
+        <Route path="/" exact component={TodoListPage} />
+        <Route path="/todo/add" component={AddTodoPage} />
+        <Route path="/todo/:slug/edit" component={EditTodoPage} />
+        <Route component={NotFoundPage} />
+      </Switch>
+    </React.Fragment>
+  );
+}
+
+export default AppPage;
