+++ src/pages/EditTodoPage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import EditTodo from '../components/EditTodo';
+
+function EditTodoPage() {
+  return (
+    <React.Fragment>
+      <p>EditTodoPage</p>
+      <EditTodo></EditTodo>
+    </React.Fragment>
+  );
+}
+
+export default EditTodoPage;
