+++ src/pages/NotFoundPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function NotFoundPage() {
+  return (
+    <p>Not found</p>
+  );
+}
+
+export default NotFoundPage;
